% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_des.r
\name{data_des}
\alias{data_des}
\title{Structure the Data as Table 1 Used in Publication.}
\usage{
data_des(
  data,
  row_var = c("vs", "mpg"),
  col_var = NULL,
  col_perc = TRUE,
  fun = c("mean", "sd"),
  comb_sym = c(" (", ")"),
  round_cate = 1,
  round_cont = 2,
  p_value = TRUE,
  test_name_cont = "t.test",
  test_name_cate = "chisq.test",
  ...
)
}
\arguments{
\item{data}{data table.}

\item{row_var}{The variable shown on the row, can't be NUll.}

\item{col_var}{The variable shown on the column, can be NULL.}

\item{col_perc}{When presented the category variable, the percentage should be calculated based on row or column. Default is column.}

\item{fun}{The function used on continuous variable.}

\item{comb_sym}{It determines how the continuous variable will be presented.}

\item{round_cate}{Decimal of category variable. Default is 1.}

\item{round_cont}{Decimal of continuous variable. Default is 2.}

\item{p_value}{If calculate the p value}

\item{test_name_cont}{test on continuous variable}

\item{test_name_cate}{test on category variable}

\item{...}{Arguments given to \code{"fun"}}
}
\description{
Category date will be presented as N(percentage). continuous data will be presented in a customized form.
}
\examples{
data(mtcars)
mtcars$judge <- TRUE
mtcars$judge[3:7] <- FALSE
mtcars$judge2 <- "TRUE"
mtcars$judge2[3:7] <- "FALSE"

# continuous varialbe was shown in mean(sd)
res_tab <- data_des(data = mtcars,
         row_var = c("vs", "mpg", "wt", "gear", "judge", "judge2"),
         fun = c("mean", "sd"),
         comb_sym = c("(", ")"),
         round_cate = 2,
         round_cont = 3)

res_tab

# continuous varialbe was shown in median[25\% quantile, 75\% quantile]
res_tab <- data_des(data = mtcars,
         row_var = c("vs", "mpg", "wt", "gear", "judge", "judge2"),
         fun = c("median", "quantile"),
         comb_sym = c("[", "]"),
         probs = c(0.25, 0.75))

res_tab

# continuous varialbe was shown in mean[25\% quantile, 75\% quantile]
res_tab <- data_des(data = mtcars,
         row_var = c("vs", "mpg", "wt", "gear", "judge", "judge2"),
         col_var = "am",
         fun = c("mean", "quantile"),
         comb_sym = c("[", "]"),
         probs = c(0.25, 0.75),
         round_cate = 2,
         round_cont = 3)

res_tab

}
\author{
Shanquan CHEN \email{shanquan0301@gmial.com}
}
\keyword{data}
\keyword{description}
