% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_data_import.r
\name{fun_data_import}
\alias{fun_data_import}
\title{Import the data from MICS}
\usage{
fun_data_import(
  country = "Pakistan",
  dataset = "hl",
  var_id = c("HH1", "HH2", "HL1"),
  var_interested_identify = c("disability", "wscore", "windex5r"),
  var_interested_match = c("ST3$", "TS5B", "TS6"),
  var_sampling_weight = "hhweight",
  level = 1
)
}
\arguments{
\item{country}{which country's data to be imported}

\item{dataset}{which data set need to be imported. Options include "hl", "ch", "fs", "wm", "mn", "hh", and others}

\item{var_id}{which variables use to generate unique id. hh is c("HH1", "HH2"), others more likely is c("HH1", "HH2", "HL1")}

\item{var_interested_identify}{identify names of variables. For example, for variable "gender", you need input "gender"}

\item{var_interested_match}{match the variables by start_with. for example, for series of variables "FCF1", "FCF2", "FCF3", FCF4", you can easily input "FCF". Then the function will extract all the variables start with "FCF"}

\item{var_sampling_weight}{sampling weight}
}
\description{
Only import data from MICS. Not for genderal use
}
\examples{
# import from hl
dat_hl <- lapply(country_name, FUN = fun_data_import,
                 dataset = "hl",
                 var_id = c("HH1", "HH2", "HL1"),
                 var_interested_identify = c("HL4", "HL6", "HH6", "ED2A", "melevel", "caretakerdis"),
                 var_interested_match = c("HL5", "HT12", "windex5", "windex10", "disability"),
                 var_sampling_weight = "hhweight")

dat_hl <- do.call("bind_rows", dat_hl)

}
\author{
Shanquan CHEN \email{shanquan0301@gmial.com}
}
\keyword{data}
\keyword{description}
