% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_var.r
\name{range_var}
\alias{range_var}
\title{Generate series of variables}
\usage{
range_var(name = "CA", num = 1:5)
}
\description{
Generate series of variables
}
\author{
Shanquan CHEN \email{shanquan0301@gmial.com}
}
