% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_comb.r
\name{reg_comb}
\alias{reg_comb}
\title{Structure the result of regression}
\usage{
reg_comb(
  reg = reg,
  data = data,
  round_p = 4,
  round_ci = 4,
  exp_transfer = FALSE,
  coef = "Estimate",
  p_value = "`Pr(>|z|)`",
  ci_low = "`2.5 \%`",
  ci_high = "`97.5 \%`",
  comb_ci = "coef(ci_low-ci_high)star, p_value"
)
}
\arguments{
\item{reg}{the regression result, like the result of lm or glm}

\item{data}{provide the data to process, only used when reg is missing}

\item{round_p}{digital for p value}

\item{round_ci}{digital for coefficient and ci}

\item{exp_transfer}{if show the exp form.}

\item{coef}{column name of coefficient}

\item{p_value}{column name of p value}

\item{ci_low}{column name of ci low}

\item{ci_high}{column name of ci high}

\item{comb_ci}{form the result to be shown}
}
\description{
Structure the result of regression
}
\examples{
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
glm.D93 <- glm(counts ~ outcome + treatment, family = poisson())
reg_comb(reg = glm.D93, exp_transfer = TRUE)

}
\author{
Shanquan CHEN \email{shanquan0301@gmial.com}
}
\keyword{data}
\keyword{description}
