% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_fun.r
\name{str_fun}
\alias{str_fun}
\title{Detect the existence of multiple strings}
\usage{
str_fun(x, str = str)
}
\arguments{
\item{x}{provide the target variable}

\item{str}{provide the strings need to be detected}
}
\description{
Detect the existence of multiple strings
}
\examples{
x <- c("counts", "outcome", "treatment")
str_fun(x, str = c("cou", "ment"))
}
\author{
Shanquan CHEN \email{shanquan0301@gmial.com}
}
\keyword{data}
\keyword{description}
