% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_join.r
\name{supp_join}
\alias{supp_join}
\title{Supplementarily join the two data sets}
\usage{
supp_join(dat_1, dat_2, full = TRUE, by = "unique_hh_ln")
}
\arguments{
\item{dat_1}{primary data table 1}

\item{dat_2}{supplement data table 2}

\item{full}{TREU keep all the cases, full_join, else left_join}

\item{by}{unique id in dat_1 and dat_2}
}
\description{
if dat_1 and dat-2 have the same variable, like gender. If gender in dat_1 is NA, then extract the gender value from dat_2
}
\examples{
dat_all <- supp_join(dat_fs, dat_hl,full = FALSE, by = "unique_hh_ln")  #merge fs and hl
}
\author{
Shanquan CHEN \email{shanquan0301@gmial.com}
}
\keyword{data}
\keyword{description}
